refactor: split handlers into separate file
